+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(crossword.r)
+
+test_check("crossword.r")